Smartmatch is experimental in 5.18 and newer perls
If you don't explicitly tell perl to stop whining, er, I mean warning
about the use of smartmatch (~~ and given/when) then every use case
generates noise.  This ugly line:

  no if $] >= 5.018, warnings => q{experimental::smartmatch};

if necessary to future-proof the code for 5.18 and newer perls.  It
also highlights the fact that it is being developed with 5,16 or older
perls, which are all sadly OEL.