Dropping the use of the experimental autoderef feature
perl5.14 introduced the ability to push to a reference, but this
feature started generating a warning in perl5.20, and appears to be
queued up for removal in the next major release of perl (which at this
point probably means 5.24).  This is purely cosmetic in the code, so
I've dropped it to future-proof this against upgrades to 5.20 and
beyond.